+++ tests/rest_tec_v1_integration/config.php
@@ -1,0 +1,4 @@
+<?php
+
+$GLOBALS['wp_tests_options'] = $GLOBALS['wp_tests_options'] ?? [];
+$GLOBALS['wp_tests_options']['permalink_structure'] = '/%postname%/';