+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### [6.15.0] 2025-08-19
+
+* Version - The Events Calendar 6.15.0 is only compatible with Event Tickets 5.26.0 and higher.
+* [EXPERIMENTAL] Feature - Introduced new REST endpoints for managing your Events, Venues and Organizers. Note: This API is for experimental use only and requires the X-TEC-EEA header. It may be subject to breaking changes in upcoming releases.
+* Language - 153 new strings added, 140 updated, 1 fuzzied, and 0 obsoleted.
+
 ### [6.14.2] 2025-07-22
 
 * Tweak - Implemented an Enable setting for Category Colors. [TEC-5533]